SDE-with-Auth0/requirements.txt
gradio
requests
beautifulsoup4
openai
nest_asyncio
Pillow
langchain
langchain-community
IPython
playwright
python-jose
auth0-python
flask
authlib
six
gunicorn
supervisor